requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
ruff
bs4
spacy
transformers
tensorflow
torch